+++ src/code/husky/Database.java
@@ -3,6 +3,7 @@ package code.husky;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 
 import org.bukkit.plugin.Plugin;
 
@@ -15,6 +16,8 @@ import org.bukkit.plugin.Plugin;
  */
 public abstract class Database {
 
+	protected Connection connection;
+	
 	/**
 	 * Plugin instance, use for plugin.getDataFolder()
 	 */
@@ -28,6 +31,7 @@ public abstract class Database {
 	 */
 	protected Database(Plugin plugin) {
 		this.plugin = plugin;
+		this.connection = null;
 	}
 
 	/**
@@ -49,14 +53,18 @@ public abstract class Database {
 	 * @throws SQLException
 	 *             if the connection cannot be checked
 	 */
-	public abstract boolean checkConnection() throws SQLException;
+	public boolean checkConnection() throws SQLException {
+		return connection != null && !connection.isClosed();
+	}
 
 	/**
 	 * Gets the connection with the database
 	 * 
 	 * @return Connection with the database, null if none
 	 */
-	public abstract Connection getConnection();
+	public Connection getConnection() {
+		return connection;
+	}
 
 	/**
 	 * Closes the connection with the database
@@ -65,7 +73,14 @@ public abstract class Database {
 	 * @throws SQLException
 	 *             if the connection cannot be closed
 	 */
-	public abstract boolean closeConnection() throws SQLException;
+	public boolean closeConnection() throws SQLException {
+		if (connection == null) {
+			return false;
+		}
+		connection.close();
+		return true;
+	}
+
 
 	/**
 	 * Executes a SQL Query<br>
@@ -80,8 +95,18 @@ public abstract class Database {
 	 * @throws ClassNotFoundException
 	 *             If the driver cannot be found; see {@link #openConnection()}
 	 */
-	public abstract ResultSet querySQL(String query) throws SQLException,
-			ClassNotFoundException;
+	public ResultSet querySQL(String query) throws SQLException,
+			ClassNotFoundException {
+		if (!checkConnection()) {
+			openConnection();
+		}
+
+		Statement statement = connection.createStatement();
+
+		ResultSet result = statement.executeQuery(query);
+
+		return result;
+	}
 
 	/**
 	 * Executes an Update SQL Query<br>
@@ -96,6 +121,16 @@ public abstract class Database {
 	 * @throws ClassNotFoundException
 	 *             If the driver cannot be found; see {@link #openConnection()}
 	 */
-	public abstract int updateSQL(String query) throws SQLException,
-			ClassNotFoundException;
+	public int updateSQL(String query) throws SQLException,
+			ClassNotFoundException {
+		if (!checkConnection()) {
+			openConnection();
+		}
+
+		Statement statement = connection.createStatement();
+
+		int result = statement.executeUpdate(query);
+
+		return result;
+	}
 }
