+++ src/code/husky/mysql/MySQL.java
@@ -2,9 +2,7 @@ package code.husky.mysql;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 
 import org.bukkit.plugin.Plugin;
 
@@ -23,7 +21,6 @@ public class MySQL extends Database {
 	private final String port;
 	private final String hostname;
 
-	private Connection connection;
 
 	/**
 	 * Creates a new MySQL instance
@@ -49,7 +46,6 @@ public class MySQL extends Database {
 		this.database = database;
 		this.user = username;
 		this.password = password;
-		this.connection = null;
 	}
 
 	@Override
@@ -64,52 +60,4 @@ public class MySQL extends Database {
 				this.user, this.password);
 		return connection;
 	}
-
-	@Override
-	public boolean checkConnection() throws SQLException {
-		return connection != null && !connection.isClosed();
-	}
-
-	@Override
-	public Connection getConnection() {
-		return connection;
-	}
-
-	@Override
-	public boolean closeConnection() throws SQLException {
-		if (connection == null) {
-			return false;
-		}
-		connection.close();
-		return true;
-	}
-
-	@Override
-	public ResultSet querySQL(String query) throws SQLException,
-			ClassNotFoundException {
-		if (!checkConnection()) {
-			openConnection();
-		}
-
-		Statement statement = connection.createStatement();
-
-		ResultSet result = statement.executeQuery(query);
-
-		return result;
-	}
-
-	@Override
-	public int updateSQL(String query) throws SQLException,
-			ClassNotFoundException {
-		if (!checkConnection()) {
-			openConnection();
-		}
-
-		Statement statement = connection.createStatement();
-
-		int result = statement.executeUpdate(query);
-
-		return result;
-	}
-
 }
