+++ src/code/husky/sqlite/SQLite.java
@@ -4,9 +4,7 @@ import java.io.File;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.logging.Level;
 
 import org.bukkit.plugin.Plugin;
@@ -21,8 +19,6 @@ import code.husky.Database;
 public class SQLite extends Database {
 	private final String dbLocation;
 
-	private Connection connection;
-
 	/**
 	 * Creates a new SQLite instance
 	 * 
@@ -34,7 +30,6 @@ public class SQLite extends Database {
 	public SQLite(Plugin plugin, String dbLocation) {
 		super(plugin);
 		this.dbLocation = dbLocation;
-		this.connection = null;
 	}
 
 	@Override
@@ -61,52 +56,5 @@ public class SQLite extends Database {
 						+ plugin.getDataFolder().toPath().toString() + "/"
 						+ dbLocation);
 		return connection;
-	}
-
-	@Override
-	public boolean checkConnection() throws SQLException {
-		return connection != null && !connection.isClosed();
-	}
-
-	@Override
-	public Connection getConnection() {
-		return connection;
-	}
-
-	@Override
-	public boolean closeConnection() throws SQLException {
-		if (connection == null) {
-			return false;
-		}
-		connection.close();
-		return true;
-	}
-
-	@Override
-	public ResultSet querySQL(String query) throws SQLException,
-			ClassNotFoundException {
-		if (!checkConnection()) {
-			openConnection();
-		}
-
-		Statement statement = connection.createStatement();
-
-		ResultSet result = statement.executeQuery(query);
-
-		return result;
-	}
-
-	@Override
-	public int updateSQL(String query) throws SQLException,
-			ClassNotFoundException {
-		if (!checkConnection()) {
-			openConnection();
-		}
-
-		Statement statement = connection.createStatement();
-
-		int result = statement.executeUpdate(query);
-
-		return result;
 	}
 }
